Rename `fine` to `fine2`
