+++ sparse_strips/vello_bench/src/fine/blend.rs
@@ -5,7 +5,7 @@ use vello_common::color::palette::css::ROYAL_BLUE;
 use vello_common::fearless_simd::Simd;
 use vello_common::paint::{Paint, PremulColor};
 use vello_common::peniko::{BlendMode, Compose, Mix};
-use vello_cpu::fine2::{Fine, FineKernel};
+use vello_cpu::fine::{Fine, FineKernel};
 use vello_dev_macros::vello_bench;
 
 pub fn blend(c: &mut Criterion) {
