+++ sparse_strips/vello_bench/src/fine/strip.rs
@@ -12,7 +12,7 @@ use vello_common::encode::EncodedPaint;
 use vello_common::fearless_simd::Simd;
 use vello_common::paint::{Paint, PremulColor};
 use vello_common::tile::Tile;
-use vello_cpu::fine2::{Fine, FineKernel};
+use vello_cpu::fine::{Fine, FineKernel};
 use vello_dev_macros::vello_bench;
 
 pub fn strip(c: &mut Criterion) {
