+++ sparse_strips/vello_cpu/src/fine/lowp/compose.rs
@@ -1,25 +1,26 @@
-use crate::fine2::Splat4thExt;
+use crate::fine::Splat4thExt;
 use crate::peniko::{BlendMode, Compose};
+use crate::util::NormalizedMulExt;
 use vello_common::fearless_simd::*;
 
 pub(crate) trait ComposeExt {
     fn compose<S: Simd>(
         &self,
         simd: S,
-        src_c: f32x16<S>,
-        bg_c: f32x16<S>,
-        alpha_mask: f32x16<S>,
-    ) -> f32x16<S>;
+        src_c: u8x32<S>,
+        bg_c: u8x32<S>,
+        alpha_mask: u8x32<S>,
+    ) -> u8x32<S>;
 }
 
 impl ComposeExt for BlendMode {
     fn compose<S: Simd>(
         &self,
         simd: S,
-        src_c: f32x16<S>,
-        bg_c: f32x16<S>,
-        alpha_mask: f32x16<S>,
-    ) -> f32x16<S> {
+        src_c: u8x32<S>,
+        bg_c: u8x32<S>,
+        alpha_mask: u8x32<S>,
+    ) -> u8x32<S> {
         match self.compose {
             Compose::SrcOver => SrcOver::compose(simd, src_c, bg_c, alpha_mask),
             Compose::Clear => Clear::compose(simd, src_c, bg_c, alpha_mask),
@@ -47,24 +48,27 @@ macro_rules! compose {
         impl $name {
             fn compose<S: Simd>(
                 simd: S,
-                src_c: f32x16<S>,
-                bg_c: f32x16<S>,
-                mask: f32x16<S>,
-            ) -> f32x16<S> {
+                src_c: u8x32<S>,
+                bg_c: u8x32<S>,
+                mask: u8x32<S>,
+            ) -> u8x32<S> {
                 let al_b = bg_c.splat_4th();
-                let al_s = src_c.splat_4th() * mask;
+                let al_s = src_c.splat_4th().normalized_mul(mask);
 
                 let fa = $fa(simd, al_s, al_b);
                 let fb = $fb(simd, al_s, al_b);
 
-                let src_c = src_c * mask;
+                let src_c = src_c.normalized_mul(mask);
 
                 if $sat {
-                    (src_c * fa + fb * bg_c)
-                        .min(f32x16::splat(simd, 1.0))
-                        .max(f32x16::splat(simd, 0.0))
+                    simd.narrow_u16x32(
+                        (simd.widen_u8x32(src_c.normalized_mul(fa))
+                            + simd.widen_u8x32(fb.normalized_mul(bg_c)))
+                        .min(u16x32::splat(simd, 255))
+                        .max(u16x32::splat(simd, 0)),
+                    )
                 } else {
-                    src_c * fa + fb * bg_c
+                    src_c.normalized_mul(fa) + fb.normalized_mul(bg_c)
                 }
             }
         }
@@ -73,79 +77,79 @@ macro_rules! compose {
 
 compose!(
     Clear,
-    |simd, _, _| f32x16::splat(simd, 0.0),
-    |simd, _, _| f32x16::splat(simd, 0.0),
+    |simd, _, _| u8x32::splat(simd, 0),
+    |simd, _, _| u8x32::splat(simd, 0),
     false
 );
 compose!(
     Copy,
-    |simd, _, _| f32x16::splat(simd, 1.0),
-    |simd, _, _| f32x16::splat(simd, 0.0),
+    |simd, _, _| u8x32::splat(simd, 255),
+    |simd, _, _| u8x32::splat(simd, 0),
     false
 );
 compose!(
     SrcOver,
-    |simd, _, _| f32x16::splat(simd, 1.0),
-    |_, al_s: f32x16<S>, _| 1.0 - al_s,
+    |simd, _, _| u8x32::splat(simd, 255),
+    |_, al_s: u8x32<S>, _| 255 - al_s,
     false
 );
 compose!(
     DestOver,
-    |_, _, al_b: f32x16<S>| 1.0 - al_b,
-    |simd, _, _| f32x16::splat(simd, 1.0),
+    |_, _, al_b: u8x32<S>| 255 - al_b,
+    |simd, _, _| u8x32::splat(simd, 255),
     false
 );
 compose!(
     Dest,
-    |simd, _, _| f32x16::splat(simd, 0.0),
-    |simd, _, _| f32x16::splat(simd, 1.0),
+    |simd, _, _| u8x32::splat(simd, 0),
+    |simd, _, _| u8x32::splat(simd, 255),
     false
 );
 compose!(
     Xor,
-    |_, _, al_b: f32x16<S>| 1.0 - al_b,
-    |_, al_s: f32x16<S>, _| 1.0 - al_s,
+    |_, _, al_b: u8x32<S>| 255 - al_b,
+    |_, al_s: u8x32<S>, _| 255 - al_s,
     false
 );
 compose!(
     SrcIn,
-    |_, _, al_b: f32x16<S>| al_b,
-    |simd, _, _| f32x16::splat(simd, 0.0),
+    |_, _, al_b: u8x32<S>| al_b,
+    |simd, _, _| u8x32::splat(simd, 0),
     false
 );
 compose!(
     DestIn,
-    |simd, _, _| f32x16::splat(simd, 0.0),
-    |_, al_s: f32x16<S>, _| al_s,
+    |simd, _, _| u8x32::splat(simd, 0),
+    |_, al_s: u8x32<S>, _| al_s,
     false
 );
 compose!(
     SrcOut,
-    |_, _, al_b: f32x16<S>| 1.0 - al_b,
-    |simd, _, _| f32x16::splat(simd, 0.0),
+    |_, _, al_b: u8x32<S>| 255 - al_b,
+    |simd, _, _| u8x32::splat(simd, 0),
     false
 );
 compose!(
     DestOut,
-    |simd, _, _| f32x16::splat(simd, 0.0),
-    |_, al_s: f32x16<S>, _| 1.0 - al_s,
+    |simd, _, _| u8x32::splat(simd, 0),
+    |_, al_s: u8x32<S>, _| 255 - al_s,
     false
 );
 compose!(
     SrcAtop,
-    |_, _, al_b: f32x16<S>| al_b,
-    |_, al_s: f32x16<S>, _| 1.0 - al_s,
+    |_, _, al_b: u8x32<S>| al_b,
+    |_, al_s: u8x32<S>, _| 255 - al_s,
     false
 );
 compose!(
     DestAtop,
-    |_, _, al_b: f32x16<S>| 1.0 - al_b,
-    |_, al_s: f32x16<S>, _| al_s,
+    |_, _, al_b: u8x32<S>| 255 - al_b,
+    |_, al_s: u8x32<S>, _| al_s,
     false
 );
 compose!(
     Plus,
-    |simd, _, _| f32x16::splat(simd, 1.0),
-    |simd, _, _| f32x16::splat(simd, 1.0),
+    |simd, _, _| u8x32::splat(simd, 255),
+    |simd, _, _| u8x32::splat(simd, 255),
     true
 );
