+++ sparse_strips/vello_cpu/src/fine/lowp/gradient.rs
@@ -45,7 +45,7 @@ impl<'a, S: Simd> Iterator for GradientFiller<'a, S> {
     }
 }
 
-impl<S: Simd> crate::fine2::Painter for GradientFiller<'_, S> {
+impl<S: Simd> crate::fine::Painter for GradientFiller<'_, S> {
     #[inline(never)]
     fn paint_u8(&mut self, buf: &mut [u8]) {
         for chunk in buf.chunks_exact_mut(64) {
