+++ sparse_strips/vello_cpu/src/fine/shaders/gradient/linear.rs
@@ -1,4 +1,4 @@
-use crate::fine2::shaders::gradient::SimdGradientKind;
+use crate::fine::shaders::gradient::SimdGradientKind;
 use core::marker::PhantomData;
 use vello_common::encode::LinearKind;
 use vello_common::fearless_simd::{Simd, f32x8};
