+++ sparse_strips/vello_cpu/src/fine/shaders/mod.rs
