+++ sparse_strips/vello_cpu/src/region.rs
@@ -3,7 +3,7 @@
 
 //! Splitting a single mutable buffer into regions that can be accessed concurrently.
 
-use crate::fine2::COLOR_COMPONENTS;
+use crate::fine::COLOR_COMPONENTS;
 use alloc::vec::Vec;
 use vello_common::coarse::WideTile;
 use vello_common::tile::Tile;
